Create FlexValueCard Component
We can use a single component to display all of the values. It can be
clicked, and we can open a modal in the future to allow its values to be
modified (when appropriate). This is a first stab at opening up space in
the middle of the screen for graphs.